Move JSONSpanExporter to seperate file
